requirements.txt
azure-storage-blob
pillow
fastapi==0.116.1
pydantic==2.11.7
python-dotenv==1.1.1
qrcode==8.2
uvicorn==0.35.0
